requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
ultralytics>=8.0.0  # YOLO
transformers>=4.30.0  # HuggingFace models
timm>=0.9.0  # Image models

# Computer Vision & AI
mediapipe>=0.10.0
face-recognition>=1.3.0
dlib>=19.24.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Data Processing
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0

# Firebase & Cloud
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0

# Web Framework (Optional)
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
psutil>=5.9.0
GPUtil>=1.4.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0

# Performance & Optimization
numba>=0.57.0
onnx>=1.14.0
onnxruntime>=1.15.0 